Refactor and fix dataclass test
