+++ tests/conftest.py
@@ -4,6 +4,45 @@ import re
 import pytest
 import responses
 
+from cognite.experimental.data_classes import AnnotationV2, AnnotationV2Filter
+
+
+@pytest.fixture
+def annotation() -> AnnotationV2:
+    return AnnotationV2(
+        annotation_type="diagrams.FileLink",
+        data={
+            "fileRef": {"id": 1, "externalId": None},
+            "pageNumber": 1,
+            "textRegion": {"xMin": 0.0, "xMax": 0.5, "yMin": 0.5, "yMax": 1.0,},
+        },
+        status="approved",
+        creating_app="UnitTest",
+        creating_app_version="0.0.1",
+        creating_user=None,
+        annotated_resource_type="file",
+        annotated_resource_id=1,
+        annotated_resource_external_id=None,
+        linked_resource_type=None,
+        linked_resource_id=None,
+        linked_resource_external_id=None,
+    )
+
+
+@pytest.fixture
+def annotation_filter() -> AnnotationV2Filter:
+    return AnnotationV2Filter(
+        annotated_resource_type="file",
+        annotated_resource_ids=[{"id": 1234}, {"external_id": "ext_1234"}],
+        annotation_type="diagrams.FileLink",
+        status="approved",
+        creating_app="UnitTest",
+        creating_user="",
+        creating_app_version="0.0.1",
+        linked_resource_type="file",
+        linked_resource_ids=[{"id": 1234}, {"external_id": "ext_1234"}],
+    )
+
 
 @pytest.fixture
 def rsps():
